Kubebuilder init of kubeconfig-operator I'm tired of manually setting up a kind cluster to act as a management cluster, only to have it broken when I delete one of the clusters or wipe out my kubeconfig.
The purpose of this controller is to help me work with the hub-spoke
model for local development. It will read the local kubeconfig, detect
local clusters (kind clusters, localstack eks clusters, etc), basically
anything that uses localhost as its entrypoint and automatically create
a namespace for that cluster and mount a usable kubeconfig secret into
that namespace.

If the cluster is deleted from the kubeconfig, or becomes unreachable,
the operator will then clean up that namespace for me.

This is basically a simple operator to help me develop and PoC
multi-cluster projects.